bpf: Promote warnings to errors and tidy code
Some of the code to fetch the userspace registers from a kernel
context were duplicated for some reason and the function wasn't used.
This wasn't caught before because we didn't make warnings errors.

Also make some other small code cleanups.

+++ src/bpf/profiler.h
@@ -205,7 +205,7 @@ unsigned long long hash_stack(native_stack_t *stack) {
 
   unsigned long long hash = seed ^ (stack->len * m);
 
-  for (int i = 0; i < MAX_STACK_DEPTH; i++) {
+  for (unsigned long long i = 0; i < MAX_STACK_DEPTH; i++) {
     // The stack is not zeroed when we initialise it, we simply
     // set the length to zero. This is a workaround to produce
     // the same hash for two stacks that might have garbage values
